gcs_to_bigquery/requirements.txt
flask
google-cloud-bigquery
google-cloud-storage
python-dotenv
gunicorn